+++ src/sanity/schemaTypes/index.ts
@@ -4,7 +4,18 @@ import {blockContentType} from './blockContentType'
 import {categoryType} from './categoryType'
 import {postType} from './postType'
 import {authorType} from './authorType'
+import {productType} from './productType'
+import {teamMemberType} from './teamMemberType'
+import {testimonialType} from './testimonialType'
 
 export const schema: { types: SchemaTypeDefinition[] } = {
-  types: [blockContentType, categoryType, postType, authorType],
+  types: [
+    blockContentType,
+    categoryType,
+    postType,
+    authorType,
+    productType,
+    teamMemberType,
+    testimonialType,
+  ],
 }